Move and update tests
