+++ py-polars/tests/unit/test_exprs.py
@@ -1034,135 +1034,3 @@ def test_is_not_deprecated() -> None:
 
     expected = pl.DataFrame({"a": [False, True, False]})
     assert_frame_equal(result, expected)
-
-
-def test_cut() -> None:
-    df = pl.DataFrame({"foo": [-2, -1, 0, 1, 2]})
-    out = df.with_columns(
-        pl.col("foo").cut([-1, 1], labels=["a", "b", "c"]).alias("cut")
-    )
-    expected = pl.DataFrame(
-        {
-            "foo": [-2, -1, 0, 1, 2],
-            "cut": pl.Series(["a", "a", "b", "b", "c"], dtype=pl.Categorical),
-        }
-    )
-    assert_frame_equal(out, expected, categorical_as_str=True)
-
-    # test with breaks
-    expected = pl.DataFrame(
-        {
-            "foo": [-2, -1, 0, 1, 2],
-            "brk": [-1.0, -1.0, 1.0, 1.0, float("inf")],
-            "foo_bin": pl.Series(
-                [
-                    "(-inf, -1]",
-                    "(-inf, -1]",
-                    "(-1, 1]",
-                    "(-1, 1]",
-                    "(1, inf]",
-                ],
-                dtype=pl.Categorical,
-            ),
-        }
-    )
-
-    # eager
-    out = df.with_columns(
-        pl.col("foo").cut([-1, 1], include_breaks=True).alias("cut")
-    ).unnest("cut")
-    assert_frame_equal(out, expected, categorical_as_str=True)
-
-    # lazy
-    out = (
-        df.lazy()
-        .with_columns(pl.col("foo").cut([-1, 1], include_breaks=True).alias("cut"))
-        .unnest("cut")
-        .collect()
-    )
-    assert_frame_equal(out, expected, categorical_as_str=True)
-
-
-def test_qcut() -> None:
-    df = pl.DataFrame({"foo": [-2, -1, 0, 1, 2]})
-
-    # pre-defined quantile probabilities
-    expected = pl.DataFrame(
-        {
-            "foo": [-2, -1, 0, 1, 2],
-            "qcut": pl.Series(["a", "a", "b", "b", "c"], dtype=pl.Categorical),
-        }
-    )
-    # eager
-    out = df.with_columns(
-        pl.col("foo").qcut([0.25, 0.75], labels=["a", "b", "c"]).alias("qcut")
-    )
-    assert_frame_equal(out, expected, categorical_as_str=True)
-    # lazy
-    out = (
-        df.lazy()
-        .with_columns(
-            pl.col("foo").qcut([0.25, 0.75], labels=["a", "b", "c"]).alias("qcut")
-        )
-        .collect()
-    )
-    assert_frame_equal(out, expected, categorical_as_str=True)
-
-    # uniform quantile probabilities
-    expected = pl.DataFrame(
-        {
-            "foo": [-2, -1, 0, 1, 2],
-            "qcut": pl.Series(
-                ["low", "low", "high", "high", "high"], dtype=pl.Categorical
-            ),
-        }
-    )
-    # eager
-    out = df.with_columns(
-        pl.col("foo").qcut(2, labels=["low", "high"], left_closed=True).alias("qcut")
-    )
-    assert_frame_equal(out, expected, categorical_as_str=True)
-    # lazy
-    out = (
-        df.lazy()
-        .with_columns(
-            pl.col("foo")
-            .qcut(2, labels=["low", "high"], left_closed=True)
-            .alias("qcut")
-        )
-        .collect()
-    )
-    assert_frame_equal(out, expected, categorical_as_str=True)
-
-    # Add both the category and the breakpoint.
-    expected = pl.DataFrame(
-        {
-            "foo": [-2, -1, 0, 1, 2],
-            "brk": [-1.0, -1.0, 1.0, 1.0, float("inf")],
-            "foo_bin": pl.Series(
-                [
-                    "(-inf, -1]",
-                    "(-inf, -1]",
-                    "(-1, 1]",
-                    "(-1, 1]",
-                    "(1, inf]",
-                ],
-                dtype=pl.Categorical,
-            ),
-        }
-    )
-    # eager
-    out = df.with_columns(
-        pl.col("foo").qcut([0.25, 0.75], include_breaks=True).alias("qcut")
-    ).unnest("qcut")
-    assert_frame_equal(out, expected, categorical_as_str=True)
-    # lazy
-    out = (
-        df.lazy()
-        .with_columns(
-            pl.col("foo").qcut([0.25, 0.75], include_breaks=True).alias("qcut")
-        )
-        .unnest("qcut")
-        .collect()
-    )
-    assert_frame_equal(out, expected, categorical_as_str=True)
